refactor(rel-import): copybutton, dashboard, datepicker updated
