+++ src/components/CopyButton/index.js
@@ -1,0 +1,1 @@
+export { CopyButton } from 'carbon-components-react';
