+++ src/components/DatePicker/index.js
@@ -1,0 +1,1 @@
+export { DatePicker, DatePickerSkeleton } from 'carbon-components-react';
