+++ src/components/DatePickerInput/index.js
@@ -1,0 +1,1 @@
+export { DatePickerInput } from 'carbon-components-react';
